+++ src/core/highlighters/date_dash.rs
@@ -85,7 +85,7 @@ impl Highlight for DateDashHighlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_date_dash_highlighter() {