+++ src/core/highlighters/date_time.rs
@@ -66,7 +66,7 @@ mod tests {
     use super::*;
     use crate::core::config::DateTimeConfig;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_time_highlighter() {