+++ src/core/highlighters/ip_v4.rs
@@ -59,7 +59,7 @@ impl Highlight for IpV4Highlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_ip_v4_highlighter() {