+++ src/core/highlighters/ip_v6.rs
@@ -58,7 +58,7 @@ impl Highlight for IpV6Highlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_ip_v6_highlighter() {