+++ src/core/highlighters/json.rs
@@ -105,7 +105,7 @@ impl Highlight for JsonHighlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_number_highlighter() {