+++ src/core/highlighters/key_value.rs
@@ -44,7 +44,7 @@ impl Highlight for KeyValueHighlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_number_highlighter() {