+++ src/core/highlighters/keyword.rs
@@ -46,7 +46,7 @@ impl Highlight for KeywordHighlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_foreground_keyword_highlighter() {