+++ src/core/highlighters/pointer.rs
@@ -93,7 +93,7 @@ fn highlight_char(c: char, number: NuStyle, x: NuStyle, letter: NuStyle) -> Stri
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_pointer_highlighter() {