+++ src/core/highlighters/unix_path.rs
@@ -69,7 +69,7 @@ impl Highlight for UnixPathHighlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_unix_path_highlighter() {