+++ src/core/highlighters/unix_process.rs
@@ -51,7 +51,7 @@ impl Highlight for UnixProcessHighlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_unix_process_highlighter() {