+++ src/core/highlighters/uuid.rs
@@ -56,7 +56,7 @@ impl Highlight for UuidHighlighter {
 mod tests {
     use super::*;
     use crate::core::tests::escape_code_converter::ConvertEscapeCodes;
-    use crate::{Color, Style};
+    use crate::style::{Color, Style};
 
     #[test]
     fn test_uuid_highlighter() {