+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.2.0
 
 - Give a warning if user is trying to read from file while also piping to `stdin`
+- Better grouping of submodules in the `tailspin` crate
 
 ## 5.1.0